fix(macos): bring TCFType into scope and cast NSInteger change count
The macOS context-capture FFI modules called TCFType methods
(as_concrete_TypeRef, as_CFTypeRef, wrap_under_create_rule) without the
trait in scope, and returned NSPasteboard::changeCount() (isize) as i64.
These never compiled because sandcastle reviewers run on Linux, where the
macos-gated modules are skipped.

+++ src-tauri/src/clipboard_context.rs
@@ -72,7 +72,7 @@ pub fn capture(app: tauri::AppHandle) {
 #[cfg(target_os = "macos")]
 pub fn platform_change_count() -> i64 {
     use objc2_app_kit::NSPasteboard;
-    NSPasteboard::generalPasteboard().changeCount()
+    NSPasteboard::generalPasteboard().changeCount() as i64
 }
 
 #[cfg(target_os = "macos")]
